Added comments, and moved the ButtonContext to a better namespace
