+++ ArkhamOverlay.Common/Events/ButtonInfoChanged.cs
@@ -1,6 +1,7 @@
 ﻿using ArkhamOverlay.Common;
 using ArkhamOverlay.Common.Enums;
 using ArkhamOverlay.Common.Services;
+using ArkhamOverlay.Common.Utils;
 using System;
 
 namespace ArkhamOverlay.Events {
