+++ ArkhamOverlay.Common/Utils/ButtonContext.cs
@@ -1,16 +1,21 @@
-﻿using System.Collections.Generic;
+﻿using ArkhamOverlay.Common.Enums;
+using System.Collections.Generic;
 using System.Linq;
 
-namespace ArkhamOverlay.Common.Enums {
+namespace ArkhamOverlay.Common.Utils {
     public enum ButtonMode { Pool, Zone }
 
+    /// <summary>
+    /// Identify a button in three parts (eventually four)
+    /// </summary>
+    /// <remarks>Eventually will need to add a Card Zone Index when we develop for multiple card zones</remarks>
     public interface IButtonContext {
         CardGroupId CardGroupId { get; }
         ButtonMode ButtonMode { get; }
         int Index { get; }
     }
 
-    public static class ButtonContextExtensions { 
+    public static class ButtonContextExtensions {
         public static bool HasSameContext(this IButtonContext a, IButtonContext b) {
             return a.CardGroupId == b.CardGroupId && a.ButtonMode == b.ButtonMode && a.Index == b.Index;
         }
@@ -30,7 +35,7 @@ namespace ArkhamOverlay.Common.Enums {
                    select potentialContext;
         }
         public static T FirstOrDefaultWithContext<T>(this IEnumerable<T> list, IButtonContext context) where T : IButtonContext {
-           return list.FirstOrDefault(x => x.HasSameContext(context));
+            return list.FirstOrDefault(x => x.HasSameContext(context));
         }
     }
 }
