+++ ArkhamOverlay/Data/CardTemplate.cs
@@ -9,6 +9,10 @@ using System.Windows.Media;
 namespace ArkhamOverlay.Data {
     public delegate void CardToggledEvent(IButton card);
 
+    /// <summary>
+    /// All information about a card, including images from either arkham db or local stored
+    /// </summary>
+    /// <remarks>Eventually this will be an abstract class so that different games can inheret from it and add game specific information here.</remarks>
     public class CardTemplate : IHasImageButton {
         public CardTemplate() {
         }
