+++ ArkhamOverlay/Data/CardZone.cs
@@ -10,6 +10,9 @@ using System.Linq;
 namespace ArkhamOverlay.Data {
     public enum CardZoneLocation { Top, Bottom }
 
+    /// <summary>
+    /// Represents a physical location in the real world (hand, act/agenda bar), and contains a list of instances of cards
+    /// </summary>
     public class CardZone {
         private readonly IEventBus _eventBus = ServiceLocator.GetService<IEventBus>();
         public CardZone(string name, CardZoneLocation location) {
